reject duplicate transformation

+++ source/common/http/path_utility.cc
@@ -1,5 +1,7 @@
 #include "common/http/path_utility.h"
 
+#include "envoy/common/exception.h"
+
 #include "common/common/logger.h"
 #include "common/http/legacy_path_canonicalizer.h"
 #include "common/runtime/runtime_features.h"
@@ -7,7 +9,6 @@
 #include "absl/strings/str_join.h"
 #include "absl/strings/str_split.h"
 #include "absl/types/optional.h"
-#include "envoy/common/exception.h"
 #include "url/url_canon.h"
 #include "url/url_canon_stdstring.h"
 
@@ -114,10 +115,11 @@ absl::optional<std::string> PathTransformer::rfcNormalize(absl::string_view orig
 
 PathTransformer::PathTransformer(envoy::type::http::v3::PathTransformation path_transformation) {
   const auto& operations = path_transformation.operations();
-  std::vector<int> operation_hashes;
+  std::vector<uint64_t> operation_hashes;
   for (auto const& operation : operations) {
     uint64_t operation_hash = MessageUtil::hash(operation);
-    if (find(operation_hashes.begin(), operation_hashes.end(), operation_hash) != operation_hashes.end()) {
+    if (find(operation_hashes.begin(), operation_hashes.end(), operation_hash) !=
+        operation_hashes.end()) {
       throw EnvoyException("Duplicate path transformation");
     }
     if (operation.has_normalize_path_rfc_3986()) {
